fix: add dynamic cacheBlockSize for InfiniteTable based on name searc… (#983)
* fix: add dynamic cacheBlockSize for InfiniteTable based on name search completion

Introduced a conditional cacheBlockSize that adjusts based on the completion of name search, enhancing performance in the SearchTreeInfinite component.

https://github.com/gisce/webclient/issues/1959

* fix: improve and avoid double calls for name search queries in infinite trees

https://github.com/gisce/webclient/issues/1959

* fix: remove useless call

* fix: introduce isNameSearchMode to optimize cacheBlockSize handling

Added a new ref to track the name search mode, allowing for a more efficient determination of cacheBlockSize in the SearchTreeInfinite component. This change improves performance by ensuring the cacheBlockSize is only set when in name search mode.

https://github.com/gisce/webclient/issues/1959

* fix: avoid duplicate call